requirements.txt
pandas==1.5.3
google-api-python-client
Unidecode
openpyxl
apache-airflow-providers-celery>=3.3.0
apache-airflow-providers-postgres
apache-airflow
kafka-python
websocket-client
pyyaml